refactor(chat): consolidate stream code into src/chat/ module
chat_protocol.rs splits into chat/event.rs (the ChatEvent protocol) and
chat/decode.rs (the StreamDecoder core); chat_ui.rs becomes
chat/render.rs. Pure move — no behavior change.
