+++ src/commands/mod.rs
@@ -6,7 +6,7 @@ pub(crate) use crate::cli::{Cli, Command};
 // Re-export the crate modules + std types the command bodies reference, so each
 // submodule can `use super::*` and keep the crate-root call style it grew up with.
 pub(crate) use crate::{
-    chat_template, chat_ui, config, dgq, flags, generate, generate_golden, golden, metal, model,
+    chat_template, config, dgq, flags, generate, generate_golden, golden, metal, model,
     sample, server, tokenizer, tools, weights,
 };
 pub(crate) use std::path::PathBuf;
